requirements.txt
asgiref==3.4.1
backports.entry-points-selectable==1.1.1
black==21.11b1
cfgv==3.3.1
click==8.0.3
colorama==0.4.4
Deprecated==1.2.13
distlib==0.3.3
Django==3.2.9
django-ninja==0.16.1
django-rq==2.5.1
filelock==3.4.0
identify==2.4.0
mypy-extensions==0.4.3
mysqlclient==2.1.0
nodeenv==1.6.0
pathspec==0.9.0
Pillow==8.4.0
platformdirs==2.4.0
pre-commit==2.15.0
pydantic==1.8.2
PyJWT==2.3.0
pytz==2021.3
PyYAML==6.0
redis==4.0.2
regex==2021.11.10
rq==1.10.0
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tomli==1.2.2
typing_extensions==4.0.0
virtualenv==20.10.0
wrapt==1.13.3
yapf==0.31.0
